+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added a flexible search filter UI for the job list. [PR #344](https://github.com/riverqueue/riverui/pull/344).
 - Added `riverlog` middleware logs to the job detail UI via a new unified attempts list, rather than separate lists for errors, attempted by, and now logs. [PR #346](https://github.com/riverqueue/riverui/pull/346).
+- Added a `RIVER_JOB_LIST_HIDE_ARGS_BY_DEFAULT` env and `ServerOpts.JobListHideArgsByDefault` setting to give the backend the ability to choose to hide args from the job list by default. This is useful if your args tend to be encoded or encrypted and not worth looking at in bulk. It can also be overridden in a new settings screen in the UI. [PR #354](https://github.com/riverqueue/riverui/pull/354).
 
 ### Changed